adding new controllers and doc. open-api.

+++ src/main/kotlin/perso/api/crypto/repository/http/paprika/model/TokenJson.kt
@@ -4,6 +4,6 @@ data class TokenJson(
     val id: String,
     val name: String,
     val rank: String,
-    val max_supply: Int,
+    val max_supply: Double,
     val quotes: QuotesJson
 )
